backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.3.1
scikit-learn==1.6.0
numpy==1.26.4
